[1731] Add recruitment cycle awareness to mcb provider show
